npm: add verify-node-gyp-called
This adds testing if node-gyp  was called by npm. It uses
npm_config_node_gyp environment variable  to inject our version of
node-gyp script to detect if it was called with either 'build' or 'rebuild'
parameter.

Also adds bufferutil, utf-8-validate, weak, hiredis and iconv packages
to the lookup.json, which utilize this new functionality.