npm: add support for custom test commands
This adds support for lookup to override test command. Provided
command will be executed by shell in the package temp directory. This
allows for creating workarounds for packages that relay on bash shell
(like ws) or ones that do not support node-dir npm parameter (like
bcrypt). It can also be used to disable running tests.